requirements.txt
pandas
PyPDF2
sentence-transformers
faiss-cpu
streamlit
google-genai
